src/test_override.py
import os
import sys
import yaml

# Add src to path
current_dir = os.path.dirname(os.path.abspath(__file__))
sys.path.append(current_dir)

from generator.llm_client import LLMClient

def test_override():
    print("=== Testing Config Override ===")
    
    # 1. Load Client
    # This will load .env internally
    config_path = os.path.join(current_dir, "../config/rag_config.yaml")
    client = LLMClient(config_path)
    
    # 2. Check what's in the config file vs environment
    provider = client.config.get("provider", "deepseek")
    conf_model = client.config.get(provider, {}).get("model")
    env_key = f"{provider.upper()}_MODEL"
    env_model = os.environ.get(env_key)
    
    print(f"Provider: {provider}")
    print(f"Model in YAML: {conf_model}")
    print(f"Model in ENV : {env_model}")
    
    if env_model:
        print(f"Expected Behavior: Use {env_model}")
    else:
        print(f"Expected Behavior: Use {conf_model}")

    # 3. Perform Generation (which triggers the logic)
    # We will hook or inspect the call, or just rely on the fact that if it works, it works.
    # To strictly verify, we can temporarily monkeypatch the client method or just trust the log.
    # Let's just run a generation. If the model ID in YAML was wrong (hypothetically) and ENV is right, success proves override.
    
    print("\nCalling Generate...")
    response = client.generate([{"role": "user", "content": "Say 'Override Works'"}])
    print(f"Response: {response}")

if __name__ == "__main__":
    test_override()
