requirements.txt
streamlit
openai
pyyaml
sentence-transformers
faiss-cpu
requests
beautifulsoup4
numpy
